Time of run was fixed.

git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@448 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocserver/internals/functions/jms_messages/triggers/message_gen.sql
@@ -29,7 +29,7 @@ begin
     for r in
         execute query
     loop
-        id_recs := id_recs || r.id;
+        id_recs := id_recs || r.id::bigint;
     end loop;
 
     id_record := id_recs[(cnt_mess+2) % cnt_rec];
